Refactored variables. Merged with main
