--- src/main/java/com/ss/ita/google/ui/pages/utils/WebElementHelper.java
@@ -1,30 +1,0 @@
-package com.ss.ita.google.ui.pages.utils;
-
-import org.openqa.selenium.JavascriptExecutor;
-import org.openqa.selenium.WebElement;
-
-import static com.ss.ita.google.ui.pages.runnerAndProperties.TestRunner.driver;
-
-public class WebElementHelper {
-    private WebElementHelper() {
-    }
-
-    public static void setAttribute(WebElement element, String attributeName, String value) {
-        String script = "arguments[0].setAttribute(arguments[1], arguments[2]);";
-        ((JavascriptExecutor) driver).executeScript(script, element, attributeName, value);
-    }
-
-    public static void removeAttribute(WebElement element, String attributeName) {
-        String script = "arguments[0].removeAttribute(arguments[1], arguments[1])";
-        ((JavascriptExecutor) driver).executeScript(script, element, attributeName);
-    }
-
-    public static void switchWebElementVisibility(WebElement element, boolean makeHidden){
-        if (makeHidden){
-            setAttribute(element, "hidden", "true");
-        }
-        else {
-            removeAttribute(element, "hidden");
-        }
-    }
-}
